backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pdf2docx==0.5.8
python-docx==1.1.0
Pillow>=10.0.0
requests>=2.31.0
apscheduler==3.10.4
